+++ CHANGELOG.md
@@ -9,6 +9,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Issue **#2264** : Users for user permissions are now retrieved from the account service plus authorisation.
 
+
+## [v7.0-beta.121] - 2021-06-02
+
+* No changes, fixing build process.
+
+
+## [v7.0-beta.120] - 2021-06-02
+
+* No changes, fixing build process.
+
+
+## [v7.0-beta.119] - 2021-06-02
+
+* Add entity relationship diagram and database DDL SQL to release artefacts.
+
 * Issue **#2241** : Changing field or operator in the query expression editor no longer deselects the selected row. 
 
 * Issue **#2241** : Made Firefox and Chrome drop downs look the same.
@@ -3467,7 +3482,10 @@ Issue **gchq/stroom-expression#22** : Add `typeOf(...)` function to dashboard.
 
 * Issue **#202** : Initial release of the new data retention policy functionality.
 
-[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-beta.118...HEAD
+[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-beta.121...HEAD
+[v7.0-beta.121]: https://github.com/gchq/stroom/compare/v7.0-beta.120...v7.0-beta.121
+[v7.0-beta.120]: https://github.com/gchq/stroom/compare/v7.0-beta.119...v7.0-beta.120
+[v7.0-beta.119]: https://github.com/gchq/stroom/compare/v7.0-beta.118...v7.0-beta.119
 [v7.0-beta.118]: https://github.com/gchq/stroom/compare/v7.0-beta.117...v7.0-beta.118
 [v7.0-beta.117]: https://github.com/gchq/stroom/compare/v7.0-beta.116...v7.0-beta.117
 [v7.0-beta.116]: https://github.com/gchq/stroom/compare/v7.0-beta.115...v7.0-beta.116